windows/requirements.txt
keyboard==0.13.5
numpy==2.2.3
pyperclip==1.9.0
python-dotenv==1.0.1
sounddevice==0.5.1
websocket-client==1.8.0
